feat: some improvements on the driver for handling using neo4j query builder
