+++ packages/neo4j/src/index.ts
@@ -6,6 +6,7 @@ export * from './Neo4jEntityManager';
 export * from './Neo4jEntityRepository';
 export * from './Neo4jSchemaGenerator';
 export * from './Neo4jQueryBuilder';
+export * from './Neo4jCypherBuilder';
 export { Neo4jEntityManager as EntityManager } from './Neo4jEntityManager';
 export { Neo4jEntityRepository as EntityRepository } from './Neo4jEntityRepository';
 export {
